+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased - 2021-xx-xx
 - Added reduced `Json` extractor with const generic size limit.
 - Added `Redirect` service.
+- Added `Acceptable` guard.
 
 # 0.1.0
 - Empty crate.